refactor: Split document tests
