+++ app/tests/shared/minimalDoc.js
@@ -1,0 +1,26 @@
+export default {
+  document: {
+    category: 'Test Report',
+    csaf_version: '2.0',
+    title: 'Minimal valid',
+    publisher: {
+      category: 'other',
+      name: 'Secvisogram Automated Tester',
+      namespace: 'https://github.com/secvisogram/secvisogram',
+    },
+    tracking: {
+      current_release_date: '2021-01-14T00:00:00.000Z',
+      id: '1',
+      initial_release_date: '2021-01-14T00:00:00.000Z',
+      revision_history: [
+        {
+          number: '1',
+          date: '2021-01-14T00:00:00.000Z',
+          summary: 'Summary',
+        },
+      ],
+      status: 'draft',
+      version: '1',
+    },
+  },
+}
